+++ src/components/form/checkable_panel/checkable_panel.test.tsx
@@ -5,6 +5,7 @@ import { requiredProps } from '../../../test/required_props';
 import { EuiCheckablePanel } from './checkable_panel';
 
 const checkablePanelRequiredProps = {
+  label: 'Label',
   id: 'id',
   onChange: () => {},
 };
@@ -13,6 +14,18 @@ describe('EuiCheckablePanel', () => {
   test('is rendered', () => {
     const component = render(
       <EuiCheckablePanel {...requiredProps} {...checkablePanelRequiredProps} />
+    );
+
+    expect(component).toMatchSnapshot();
+  });
+
+  test('renders a checkbox when specified', () => {
+    const component = render(
+      <EuiCheckablePanel
+        {...requiredProps}
+        {...checkablePanelRequiredProps}
+        checkableType="checkbox"
+      />
     );
 
     expect(component).toMatchSnapshot();