+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added `disabled` prop to the `EuiCheckboxGroup` definition ([#2545](https://github.com/elastic/eui/pull/2545))
 - Added `disabled` option to the `option` attribute of the `options` object that is passed to the `EuiCheckboxGroup` so that checkboxes in a group can be individually disabled ([#2548](https://github.com/elastic/eui/pull/2548))
 - Added `EuiAspectRatio` component that allows for responsively resizing embeds ([#2535](https://github.com/elastic/eui/pull/2535))
-- Added `CheckablePanel` component, for radio buttons or checkboxes with complex child content ([#???](https://github.com/elastic/eui/pull/???))
+- Added `EuiCheckablePanel` component, for radio buttons or checkboxes with complex child content; updated `EuiCheckbox` and `EuiCheckboxGroup` to TypeScript. ([#2555](https://github.com/elastic/eui/pull/2555))
 
 **Bug fixes**